Реорганизация структуры store
