--- src/store/actions/data.js
@@ -1,39 +1,0 @@
-export const ActionType = {
-  LOAD_FILMS: `LOAD_FILMS`,
-  LOAD_FILM: `LOAD_FILM`,
-  LOAD_REVIEWS: `LOAD_REWIEVS`,
-  LOAD_PROMO_FILM: `LOAD_PROMO_FILM`,
-  LOAD_FAVORITE_FILMS: `LOAD_FAVORITE_FILMS`,
-  UPDATE_IS_MOVIE_FAVOTIRE: `UPDATE_IS_MOVIE_FAVOTIRE`,
-  UPDATE_FILMS_INFO: `UPDATE_FILMS_INFO`,
-};
-
-export const loadFilms = (films) => ({
-  type: ActionType.LOAD_FILMS,
-  payload: films,
-});
-
-export const loadFilm = (film) => ({
-  type: ActionType.LOAD_FILM,
-  payload: film,
-});
-
-export const loadReviews = (reviews) => ({
-  type: ActionType.LOAD_REVIEWS,
-  payload: reviews,
-});
-
-export const loadPromoFilm = (film) => ({
-  type: ActionType.LOAD_PROMO_FILM,
-  payload: film,
-});
-
-export const loadFavoriteFilms = (films) => ({
-  type: ActionType.LOAD_FAVORITE_FILMS,
-  payload: films,
-});
-
-export const updateFilmsInfo = (film)=>({
-  type: ActionType.UPDATE_FILMS_INFO,
-  payload: film,
-});
