+++ src/store/catalog/catalog.js
@@ -1,12 +1,18 @@
 import {ALL_GENRES_FILTER, SHOWN_FILMS_INITIAL_COUNT} from "../../const";
 import {extend} from "../../utils/common";
-import {ActionType} from "../actions/catalog";
+
+export const ActionType = {
+  CHANGE_CURRENT_GENRE: `CHANGE_CURRENT_GENRE`,
+  INCREASE_SHOWN_FILMS_COUNT: `INCREASE_SHOWN_FILMS_COUNT`,
+  RESET_SHOWN_FILMS_COUNT: `RESET_SHOWN_FILMS_COUNT`,
+};
 
 const initialState = {
   currentGenre: ALL_GENRES_FILTER,
   shownFilmsCount: SHOWN_FILMS_INITIAL_COUNT,
-
 };
+
+// Reducer
 
 export const catalog = (state = initialState, action) => {
   switch (action.type) {
@@ -23,3 +29,20 @@ export const catalog = (state = initialState, action) => {
 
   return state;
 };
+
+// Actions
+
+export const changeCurrentGenre = (genre)=>({
+  type: ActionType.CHANGE_CURRENT_GENRE,
+  payload: genre
+});
+
+export const increaseShownFilmsCount = (step)=>({
+  type: ActionType.INCREASE_SHOWN_FILMS_COUNT,
+  payload: step
+});
+
+export const resetShownFilmsCount = ()=>({
+  type: ActionType.RESET_SHOWN_FILMS_COUNT,
+  payload: {}
+});
