+++ src/store/user/user.js
@@ -1,6 +1,13 @@
-import {AuthorizationStatus, RequestStatus} from "../../const";
 import {extend} from "../../utils/common";
-import {ActionType} from "../actions/user";
+import {RequestStatus, AuthorizationStatus} from "../../const";
+
+export const ActionType = {
+  SET_AUTHORIZATION_STATUS: `SET_AUTHORIZATION_STATUS`,
+  LOAD_USER: `LOAD_USER`,
+  USER_REQUESTED: `USER_REQUESTED`,
+  USER_RECEIVED: `USER_RECEIVED`,
+  USER_REQUEST_FAILED: `USER_REQUEST_FAILED`,
+};
 
 const initialState = {
   // TODO: вернуть состояние по умолчанию AuthorizationStatus.NO_AUTH
@@ -10,7 +17,9 @@ const initialState = {
   errorCode: 0,
 };
 
-const user = (state = initialState, action) => {
+// Reducer
+
+export const user = (state = initialState, action) => {
   switch (action.type) {
     case ActionType.SET_AUTHORIZATION_STATUS:
       return extend(state, {authorizationStatus: action.payload});
@@ -31,4 +40,31 @@ const user = (state = initialState, action) => {
   return state;
 };
 
-export {user};
+
+// Actions
+
+export const setAuthorizationStatus = (status) => ({
+  type: ActionType.SET_AUTHORIZATION_STATUS,
+  payload: status,
+});
+
+export const loadUser = (userInfo) => ({
+  type: ActionType.LOAD_USER,
+  payload: userInfo,
+});
+
+export const userRequested = () => ({
+  type: ActionType.USER_REQUESTED,
+  payload: {},
+});
+
+export const userReceived = () => ({
+  type: ActionType.USER_RECEIVED,
+  payload: {},
+});
+
+export const userRequestFailed = (error) => ({
+  type: ActionType.USER_REQUEST_FAILED,
+  payload: error,
+});
+
